# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/generate": {
            "post": {
                "description": "This route generates a new secret have the user's data",
                "consumes": [
                    "application/json",
                    "text/xml"
                ],
                "produces": [
                    "application/json",
                    "text/xml"
                ],
                "tags": [
                    "token"
                ],
                "summary": "Create a new secret",
                "parameters": [
                    {
                        "description": "Generate a secret",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Data"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ResultToken"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorModel"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorModel"
                        }
                    }
                }
            }
        },
        "/get/{token}": {
            "post": {
                "description": "This routes generate new secret have the user's data",
                "consumes": [
                    "application/json",
                    "text/xml"
                ],
                "produces": [
                    "application/json",
                    "text/xml"
                ],
                "tags": [
                    "token"
                ],
                "summary": "analyze the secret",
                "parameters": [
                    {
                        "type": "string",
                        "description": "get the secret info",
                        "name": "token",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorModel"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorModel"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Data": {
            "type": "object",
            "required": [
                "data",
                "expire",
                "views"
            ],
            "properties": {
                "data": {
                    "type": "string"
                },
                "expire": {
                    "type": "integer"
                },
                "views": {
                    "type": "integer"
                }
            }
        },
        "models.ErrorModel": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "models.ResponseData": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "object": {
                    "$ref": "#/definitions/models.Secret"
                }
            }
        },
        "models.ResultToken": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "models.Secret": {
            "type": "object",
            "required": [
                "expireDate",
                "views"
            ],
            "properties": {
                "expireDate": {
                    "type": "string"
                },
                "views": {
                    "type": "integer"
                }
            }
        }
    }
}